+++ Changes.md
@@ -1,6 +1,12 @@
 ### Unreleased
 
 
+### [1.1.0] - 2022-07-20
+
+- fix: timeout DNS 1 second before plugin timeout
+- resolve_ptr_names: simplify with async/await & Promise.all
+
+
 ### [1.0.6] - 2022-06-05
 
 - ci: update GHA shared workflows
@@ -45,4 +51,7 @@
 ### 1.0.0 - 2017-02-05
 
 - initial release
+
+
 [1.0.6]: https://github.com/haraka/haraka-plugin-fcrdns/releases/tag/1.0.6
+[1.1.0]: https://github.com/haraka/haraka-plugin-fcrdns/releases/tag/1.1.0